pitching_pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pitching-pipeline"
version = "0.1.0"
description = "7-view 240Hz 3DGS-based pitching biomechanics analysis pipeline"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.11",
    "opencv-python>=4.8",
    "torch>=2.0",
    "smplx>=0.1.28",
    "pyyaml>=6.0",
    "matplotlib>=3.7",
    "tqdm",
]

[project.optional-dependencies]
full = [
    "opensim>=4.5",
    "c3d",
    "pingouin>=0.5.3",
    "trimesh>=3.23",
]

[project.scripts]
pitching-pipeline = "scripts.run_pipeline:main"
